Split out assets, collections and items from importer.tasks
